refactor: move Popup in 'features' and popup components in 'ui/popups'
